requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.4
django-allauth==65.3.0
djangorestframework==3.15.2
psycopg2==2.9.10
pycparser==2.22
sqlparse==0.5.3
tzdata==2024.2

#pip install -r requirements.txt